% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{age_labels}
\alias{age_labels}
\title{Provides age_labels that follows the mg standard}
\usage{
age_labels(age_cuts)
}
\arguments{
\item{age_cuts}{The lower bound of the groups (0 is implicitly included)}
}
\value{
A vector of labels with zero-padded numerics so they can be sorted easily
}
\description{
Provides age_labels that follows the mg standard
}
\examples{
age_labels(c(5, 12, 20, 30))
}
