% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{aggregate_age}
\alias{aggregate_age}
\title{Groups age (or other numerics) into a bins set by age_cuts}
\usage{
aggregate_age(age, age_cuts)
}
\arguments{
\item{age}{The vector of numbers to group}

\item{age_cuts}{The lower bound of the groups (0 is implicitly included)}
}
\value{
A factor of same length as age, but containing the name of the bin
}
\description{
Groups age (or other numerics) into a bins set by age_cuts
}
