% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{aggregate_age_sql}
\alias{aggregate_age_sql}
\title{Groups age (or other numerics) into a bins set by age_cuts (builds query for database table object)}
\usage{
aggregate_age_sql(age_column, age_cuts)
}
\arguments{
\item{age_column}{The name of the column with the age group}

\item{age_cuts}{The lower bound of the groups (0 is implicitly included)}
}
\value{
A query to bin the age_column. Use in combination with "mutate" and "!!"
}
\description{
Groups age (or other numerics) into a bins set by age_cuts (builds query for database table object)
}
\examples{
conn <- get_connection()

get_table(conn, "prod.covid_19_maalgruppe") \%>\%
     mutate(age_group = !!aggregate_age_sql(vaccineage, age_cuts = c(30, 60)))

close_connection(conn)
}
