% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{assert_dbtable_like}
\alias{assert_dbtable_like}
\title{checkmate helper: Assert for "generic" db_table type}
\usage{
assert_dbtable_like(db_table, ..., add = NULL)
}
\arguments{
\item{db_table}{Object to test if is of class "tbl_dbi" or character on form "schema.table"}

\item{...}{Parameters passed to checkmate::check_*}

\item{add}{\code{AssertCollection} to add assertions to}
}
\description{
checkmate helper: Assert for "generic" db_table type
}
