% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{assert_id_like}
\alias{assert_id_like}
\title{checkmate helper: Assert for "generic" id structure}
\usage{
assert_id_like(id, ..., add = NULL)
}
\arguments{
\item{id}{Object to test if is of class "Id" or character on form "schema.table"}

\item{...}{Parameters passed to checkmate::check_*}

\item{add}{\code{AssertCollection} to add assertions to}
}
\value{
Throws an error (or pushes the error message to an 
  \code{\link[checkmate]{AssertCollection}} if \code{add} is not \code{NULL}) 
  if the checks fail and invisibly returns \code{TRUE} otherwise.
}
\description{
checkmate helper: Assert for "generic" id structure
}
