% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{assert_timestamp_like}
\alias{assert_timestamp_like}
\title{checkmate helper: Assert "generic" timestamp type}
\usage{
assert_timestamp_like(timestamp, ..., add = NULL)
}
\arguments{
\item{timestamp}{Object to test if is POSIX or character}

\item{...}{parameters passed to checkmate::check_*}

\item{add}{\code{AssertCollection} to add assertions to}
}
\value{
Throws an error (or pushes the error message to an 
  \code{\link[checkmate]{AssertCollection}} if \code{add} is not \code{NULL}) 
  if the checks fail and invisibly returns \code{TRUE} otherwise.
}
\description{
checkmate helper: Assert "generic" timestamp type
}
