% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R, R/db_accessing_functions.R
\name{close_connection}
\alias{close_connection}
\title{Close connection to the DB}
\usage{
close_connection(conn)

close_connection(conn)
}
\arguments{
\item{conn}{An object that inherits from DBIConnection (as generated by get_connection())'}
}
\description{
Close connection to the DB

Close connection to the DB
}
