% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_timestamps.R
\name{db_timestamp}
\alias{db_timestamp}
\title{Translate timestamps to formats the database supports}
\usage{
db_timestamp(timestamp, conn = NULL)
}
\arguments{
\item{timestamp}{The timestamp to be transformed to the DB type. Can be character.}

\item{conn}{A \code{DBIConnection} to the DB where the timestamp should be stored.}
}
\value{
The given timestamp converted to a SQL-backend dependent timestamp.
}
\description{
Translate timestamps to formats the database supports
}
\examples{
conn <- get_connection(drv = RSQLite::SQLite())

db_timestamp(Sys.time(), conn)

close_connection(conn)
}
