% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend_oracle.R
\name{dbQuoteIdentifier,JDBCConnection,character-method}
\alias{dbQuoteIdentifier,JDBCConnection,character-method}
\title{Quote identifiers for Oracle}
\usage{
\S4method{dbQuoteIdentifier}{JDBCConnection,character}(conn, x, ...)
}
\arguments{
\item{conn}{A JDBCConnection}

\item{x}{Character vector of identifiers to quote}

\item{...}{Additional arguments (ignored)}
}
\value{
SQL object with quoted identifiers
}
\description{
Quote identifiers for Oracle
}
