% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{defer_db_cleanup}
\alias{defer_db_cleanup}
\title{Delete table at function exit}
\usage{
defer_db_cleanup(tbl_sql)
}
\arguments{
\item{tbl_sql}{A "bare" reference to a sql table}
}
\value{
NULL (called for side effects)
}
\description{
This function marks a table for deletion once the current function exits.
}
\examples{
conn <- get_connection(drv = RSQLite::SQLite())

mt <- dplyr::copy_to(conn, mtcars)
id_mt <- id(mt)

defer_db_cleanup(mt)

DBI::dbExistsTable(conn, id_mt) # TRUE

withr::deferred_run()

DBI::dbExistsTable(conn, id_mt) # FALSE

close_connection(conn)
}
