% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/digest_to_checksum.R
\name{digest_internal}
\alias{digest_internal}
\alias{digest_to_checksum_internal}
\alias{digest_to_checksum_internal.default}
\alias{digest_to_checksum_native_md5}
\alias{digest_to_checksum_internal.tbl_PqConnection}
\alias{digest_to_checksum_internal.data.frame}
\alias{digest_to_checksum_internal.tibble}
\title{Some backends have native md5 support, these use this function}
\usage{
digest_to_checksum_internal(.data, col)

\method{digest_to_checksum_internal}{default}(.data, col)

digest_to_checksum_native_md5(.data, col)

\method{digest_to_checksum_internal}{tbl_PqConnection}(.data, col)

\method{digest_to_checksum_internal}{data.frame}(.data, col)

\method{digest_to_checksum_internal}{tibble}(.data, col)
}
\arguments{
\item{.data}{A data frame, a tibble, a data.table or a tbl.}

\item{col}{The name of column the checksums will be placed in}
}
\description{
Some backends have native md5 support, these use this function
}
