% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid_functions.R
\name{focus_lineage_map}
\alias{focus_lineage_map}
\title{Groups known pango-lineages by the major variants provided}
\usage{
focus_lineage_map(pangolin_data, variant_major = c("BA.2", "BA.5"))
}
\arguments{
\item{pangolin_data}{A lazy-query of pangolin data on which to build the map}

\item{variant_major}{A list of lineages to subdivide variants by. Lineages are matched to the most specific
variant_major provided. Default is BA.2 and BA.5}
}
\value{
A lazy query that contains lineage and focus_linage (on the form "variant: sub-lineage")
}
\description{
Groups known pango-lineages by the major variants provided
}
\examples{
conn <- get_connection()

pangolin <- get_table(conn, "mg.lineage_info")

focus_lineage_map(pangolin) # Defaults to "BA.2", "BA.5"
focus_lineage_map(pangolin, c("BA.2", "BA.2.12", "BA.5"))

close_connection(conn)
}
