% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printing_helpers.R
\name{format_danish}
\alias{format_danish}
\title{print numbers with Danish thousands separator and decimal marks}
\usage{
format_danish(n, fmt = "f")
}
\arguments{
\item{n}{Number to format}

\item{fmt}{A printf-like format specifier}
}
\description{
print numbers with Danish thousands separator and decimal marks
}
