% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printing_helpers.R
\name{format_timestamp}
\alias{format_timestamp}
\title{formats timestamps as 2022-13-09 09:30,213}
\usage{
format_timestamp(ts)
}
\arguments{
\item{ts}{Timestamp to format}
}
\description{
formats timestamps as 2022-13-09 09:30,213
}
