% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_catalog.R
\name{get_catalog}
\alias{get_catalog}
\alias{get_catalog.Microsoft SQL Server}
\title{Get the current catalog of a database-related objects}
\usage{
get_catalog(obj, ...)

\method{get_catalog}{`Microsoft SQL Server`}(obj, temporary = FALSE, ...)
}
\arguments{
\item{obj}{The object from which to retrieve a schema}

\item{...}{Further arguments passed to methods.}

\item{temporary}{(\code{logical(1)}) \cr
Should the reference be to the temporary schema/database?}
}
\value{
For \verb{get_catalog.Microsoft SQL Server}, the current database context of the connection or "tempdb" if
\code{temporary = TRUE}.

For \verb{get_catalog.*}, \code{NULL} is returned.
}
\description{
Get the current catalog of a database-related objects
}
\examples{
conn <- get_connection(drv = RSQLite::SQLite())

dplyr::copy_to(conn, mtcars, name = "mtcars", temporary = FALSE)

get_catalog(conn)
get_catalog(get_table(conn, id("mtcars", conn = conn)))

close_connection(conn)
}
