% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_table.R
\name{get_tables}
\alias{get_tables}
\title{Gets the available tables}
\usage{
get_tables(conn, pattern = NULL, show_temporary = TRUE)
}
\arguments{
\item{conn}{An object that inherits from DBIConnection (as generated by get_connection())'}

\item{pattern}{A regex pattern with which to subset the returned tables}

\item{show_temporary}{(\code{logical})\cr
Should temporary tables be listed?}
}
\value{
A data.frame containing table names in the DB
}
\description{
Gets the available tables
}
\examples{
conn <- get_connection(drv = RSQLite::SQLite())

dplyr::copy_to(conn, mtcars, name = "my_test_table_1", temporary = FALSE)
dplyr::copy_to(conn, mtcars, name = "my_test_table_2")

get_tables(conn, pattern = "my_[th]est")
get_tables(conn, pattern = "my_[th]est", show_temporary = FALSE)

close_connection(conn)
}
