% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{id}
\alias{id}
\title{Convenience function for DBI::Id}
\usage{
id(db_table_id, conn = NULL, allow_table_only = TRUE)
}
\arguments{
\item{db_table_id}{A \code{\link[DBI:Id]{DBI::Id()}} object or a character string readable by \link{id}.}

\item{conn}{An object that inherits from DBIConnection (as generated by get_connection())'}

\item{allow_table_only}{logical. If \code{TRUE}, allows for returning an \code{DBI::Id} with \code{table} = \code{myschema.table} if schema \code{myschema}
is not found in \code{conn}.}
}
\value{
A DBI::Id object parsed from db_table_id
}
\description{
Convenience function for DBI::Id
}
\details{
The given db_table_id is parsed using an assumption of "schema.table" syntax into
a DBI::Id object with corresponding schema (if the conn supports it) and table values.
}
\examples{
id("schema.table")
}
\seealso{
\link[DBI:Id]{DBI::Id} which this function wraps.
}
