% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_joins.R
\name{join_warn}
\alias{join_warn}
\title{A warning to users that SQL does not match on NA by default}
\usage{
join_warn()
}
\value{
A warning that *_joins on SQL backends does not match NA by default
}
\description{
A warning to users that SQL does not match on NA by default
}
