% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_joins.R
\name{join_warn_experimental}
\alias{join_warn_experimental}
\title{A warning to users that SQL does not match on NA by default}
\usage{
join_warn_experimental()
}
\value{
A warning that *_joins are still experimental
}
\description{
A warning to users that SQL does not match on NA by default
}
