% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_joins.R
\name{mg_right_join}
\alias{mg_right_join}
\title{dbplyr and SQLite does not work right now for right_joins it seems
so we "fix" it by doing a left join on SQLiteConnections}
\usage{
mg_right_join(x, y, sql_on, renamer, ...)
}
\arguments{
\item{x, y}{A pair of lazy data frames backed by database queries.}

\item{sql_on}{A custom join predicate as an SQL expression.
Usually joins use column equality, but you can perform more complex
queries by supply \code{sql_on} which should be a SQL expression that
uses \code{LHS} and \code{RHS} aliases to refer to the left-hand side or
right-hand side of the join respectively.}

\item{renamer}{named list generated by select_na_sql}

\item{...}{Other parameters passed onto methods.}
}
\description{
dbplyr and SQLite does not work right now for right_joins it seems
so we "fix" it by doing a left join on SQLiteConnections
}
