% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{nrow}
\alias{nrow}
\title{nrow() but also works on remote tables}
\usage{
nrow(.data)
}
\arguments{
\item{.data}{lazy_query to parse}
}
\value{
The number of records in the object
}
\description{
nrow() but also works on remote tables
}
\examples{
conn <- get_connection()

get_table(conn, "mg.epicpr") \%>\%
     nrow()

close_connection(conn)
}
