% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Logger.R
\name{NullLogger}
\alias{NullLogger}
\title{No-logging Logger}
\value{
A new instance of the \code{Logger} \link[R6:R6Class]{R6} class where no logging will be done.
}
\description{
Create a \link{Logger} object that does no logging.
}
\examples{
logger <- NullLogger$new()
}
\section{Super class}{
\code{\link[SCDB:Logger]{SCDB::Logger}} -> \code{NullLogger}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-NullLogger-new}{\code{NullLogger$new()}}
\item \href{#method-NullLogger-log_info}{\code{NullLogger$log_info()}}
\item \href{#method-NullLogger-log_warn}{\code{NullLogger$log_warn()}}
\item \href{#method-NullLogger-log_error}{\code{NullLogger$log_error()}}
\item \href{#method-NullLogger-log_to_db}{\code{NullLogger$log_to_db()}}
\item \href{#method-NullLogger-finalize_db_entry}{\code{NullLogger$finalize_db_entry()}}
\item \href{#method-NullLogger-clone}{\code{NullLogger$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="SCDB" data-topic="Logger" data-id="finalize"><a href='../../SCDB/html/Logger.html#method-Logger-finalize'><code>SCDB::Logger$finalize()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NullLogger-new"></a>}}
\if{latex}{\out{\hypertarget{method-NullLogger-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{NullLogger} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NullLogger$new(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Captures arguments given, but does nothing}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NullLogger-log_info"></a>}}
\if{latex}{\out{\hypertarget{method-NullLogger-log_info}{}}}
\subsection{Method \code{log_info()}}{
Matches the signature of \code{Logger$log_info()}, but does nothing.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NullLogger$log_info(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Captures arguments given, but does nothing}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NullLogger-log_warn"></a>}}
\if{latex}{\out{\hypertarget{method-NullLogger-log_warn}{}}}
\subsection{Method \code{log_warn()}}{
Matches the signature of \code{Logger$log_warn()}, but does nothing.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NullLogger$log_warn(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Captures arguments given, but does nothing}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NullLogger-log_error"></a>}}
\if{latex}{\out{\hypertarget{method-NullLogger-log_error}{}}}
\subsection{Method \code{log_error()}}{
Matches the signature of \code{Logger$log_error()}, but does nothing.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NullLogger$log_error(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Captures arguments given, but does nothing}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NullLogger-log_to_db"></a>}}
\if{latex}{\out{\hypertarget{method-NullLogger-log_to_db}{}}}
\subsection{Method \code{log_to_db()}}{
Matches the signature of \code{Logger$log_to_db()}, but does nothing.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NullLogger$log_to_db(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Captures arguments given, but does nothing}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NullLogger-finalize_db_entry"></a>}}
\if{latex}{\out{\hypertarget{method-NullLogger-finalize_db_entry}{}}}
\subsection{Method \code{finalize_db_entry()}}{
Matches the signature of \code{Logger$finalize_db_entry()}, but does nothing.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NullLogger$finalize_db_entry(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Captures arguments given, but does nothing}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NullLogger-clone"></a>}}
\if{latex}{\out{\hypertarget{method-NullLogger-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NullLogger$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
