% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printing_helpers.R
\name{printr}
\alias{printr}
\title{cat printing with default new line}
\usage{
printr(..., file = nullfile(), sep = "")
}
\arguments{
\item{...}{The normal input to cat}

\item{file}{Path of an output file to append the output to}

\item{sep}{The separator given to cat}
}
\description{
cat printing with default new line
}
