% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_schema.R
\name{schema_exists}
\alias{schema_exists}
\title{Test if a schema exists in given connection}
\usage{
schema_exists(conn, schema)
}
\arguments{
\item{conn}{An object that inherits from DBIConnection (as generated by get_connection())'}

\item{schema}{A character string giving the schema name}
}
\description{
Test if a schema exists in given connection
}
\examples{

conn <- get_connection(drv = RSQLite::SQLite())
schema_exists(conn, "test")

}
