% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{assert_data_like}
\alias{assert_data_like}
\title{checkmate helper: Assert "generic" data.table/data.frame/tbl/tibble type}
\usage{
assert_data_like(.data, ..., add = NULL)
}
\arguments{
\item{.data}{Object to test if is data.table, data.frame, tbl or tibble}

\item{...}{Parameters passed to checkmate::check_*}

\item{add}{\code{AssertCollection} to add assertions to}
}
\value{
Throws an error (or pushes the error message to an 
  \code{\link[checkmate]{AssertCollection}} if \code{add} is not \code{NULL}) 
  if the checks fail and invisibly returns \code{TRUE} otherwise.
}
\description{
checkmate helper: Assert "generic" data.table/data.frame/tbl/tibble type
}
