% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{assert_dbtable_like}
\alias{assert_dbtable_like}
\title{checkmate helper: Assert for "generic" db_table type}
\usage{
assert_dbtable_like(db_table, ..., add = NULL)
}
\arguments{
\item{db_table}{Object to test if is of class "tbl_dbi" or character on form "schema.table"}

\item{...}{Parameters passed to checkmate::check_*}

\item{add}{\code{AssertCollection} to add assertions to}
}
\value{
Throws an error (or pushes the error message to an 
  \code{\link[checkmate]{AssertCollection}} if \code{add} is not \code{NULL}) 
  if the checks fail and invisibly returns \code{TRUE} otherwise.
}
\description{
checkmate helper: Assert for "generic" db_table type
}
