% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{assert_id_like}
\alias{assert_id_like}
\title{checkmate helper: Assert for "generic" id structure}
\usage{
assert_id_like(id, ..., add = NULL)
}
\arguments{
\item{id}{Object to test if is of class "Id" or character on form "schema.table"}

\item{...}{Parameters passed to checkmate::check_*}

\item{add}{\code{AssertCollection} to add assertions to}
}
\description{
checkmate helper: Assert for "generic" id structure
}
