% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{assert_timestamp_like}
\alias{assert_timestamp_like}
\title{checkmate helper: Assert "generic" timestamp type}
\usage{
assert_timestamp_like(timestamp, ..., add = NULL)
}
\arguments{
\item{timestamp}{Object to test if is POSIX or character}

\item{...}{parameters passed to checkmate::check_*}

\item{add}{\code{AssertCollection} to add assertions to}
}
\description{
checkmate helper: Assert "generic" timestamp type
}
