% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{close_connection}
\alias{close_connection}
\title{Close connection to the DB}
\usage{
close_connection(conn)
}
\arguments{
\item{conn}{(\code{DBIConnection(1)})\cr
Connection object.}
}
\value{
\code{dbDisconnect()} returns \code{TRUE}, invisibly.
}
\description{
Close connection to the DB
}
\examples{
  conn <- get_connection(drv = RSQLite::SQLite())

  close_connection(conn)
}
