% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_table.R
\name{create_index}
\alias{create_index}
\title{Create the indexes on table}
\usage{
create_index(conn, db_table_id, columns)
}
\arguments{
\item{conn}{(\code{DBIConnection})\cr
A connection to a database.}

\item{db_table_id}{(\code{Id})\cr
The table to create the index for.}

\item{columns}{(\code{character()})\cr
The columns that should be unique.}
}
\value{
NULL (called for side effects)
}
\description{
Create the indexes on table
}
