% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_timestamps.R
\name{db_timestamp}
\alias{db_timestamp}
\alias{db_timestamp.default}
\alias{db_timestamp.SQLiteConnection}
\title{Determine the type of timestamps the DB supports}
\usage{
db_timestamp(timestamp, conn = NULL)

\method{db_timestamp}{default}(timestamp, conn)

\method{db_timestamp}{SQLiteConnection}(timestamp, conn)
}
\arguments{
\item{timestamp}{(\code{POSIXct(1)} or \code{character(1)})\cr
The timestamp to be transformed to the DB type.}

\item{conn}{(\code{DBIConnection(1)})\cr
Connection object.}
}
\value{
The given timestamp converted to a SQL-backend dependent timestamp.
conn <- get_connection(drv = RSQLite::SQLite())

db_timestamp(Sys.time(), conn)

close_connection(conn)
}
\description{
Determine the type of timestamps the DB supports
}
