% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_accessing_functions.R, R/get_table.R
\name{get_tables}
\alias{get_tables}
\title{Gets the available tables}
\usage{
get_tables(conn, pattern = NULL)

get_tables(conn, pattern = NULL)
}
\arguments{
\item{conn}{An object that inherits from DBIConnection (as generated by get_connection())'}

\item{pattern}{A regex pattern with which to subset the returned tables}
}
\value{
A data.frame containing table names in the DB

A data.frame containing table names in the DB
}
\description{
Gets the available tables

Gets the available tables
}
\examples{
conn <- get_connection()

get_tables(conn)

close_connection(conn)
conn <- get_connection()

get_tables(conn)

close_connection(conn)
}
