% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{\%notin\%}
\alias{\%notin\%}
\title{not-in operator}
\usage{
x \%notin\% table
}
\arguments{
\item{x}{vector or \code{NULL}: the values to be matched.
    \link[base]{Long vectors} are supported.}

\item{table}{vector or \code{NULL}: the values to be matched against.
    \link[base]{Long vectors} are not supported.}
}
\value{
A logical vector, indicating if a match was NOT located for each element of x
}
\description{
not-in operator
}
\examples{
2 \%notin\% c(1,3) # TRUE
}
