% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R, R/db_accessing_functions.R
\name{id}
\alias{id}
\title{Convenience function for DBI::Id}
\usage{
id(db_table_id, conn = NULL)

id(db_table_id, conn = NULL)
}
\arguments{
\item{db_table_id}{A \code{\link[DBI:Id]{DBI::Id()}} object or a character string readable by \link{id}.}

\item{conn}{An object that inherits from DBIConnection (as generated by get_connection())'}
}
\description{
Convenience function for DBI::Id

Convenience function for DBI::Id
}
\seealso{
\link[DBI:Id]{DBI::Id} which this function wraps.

\link[DBI:Id]{DBI::Id} which this function wraps.
}
