% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{nrow}
\alias{nrow}
\title{nrow() but also works on remote tables}
\usage{
nrow(.data)
}
\arguments{
\item{.data}{(\code{data.frame(1)}, \code{tibble(1)}, \code{data.table(1)}, or \code{tbl_dbi(1)})\cr
Data object.}
}
\value{
The number of records in the object.
conn <- get_connection(drv = RSQLite::SQLite())

m <- dplyr::copy_to(conn, mtcars)
nrow(m) == nrow(mtcars) # TRUE

close_connection(conn)
}
\description{
nrow() but also works on remote tables
}
