% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mg-package.R
\docType{package}
\name{SCDB-package}
\alias{SCDB}
\alias{SCDB-package}
\title{SCDB: Easily access and maintain time-based versioned data}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

This package contains scripts for easily accessing and updating a database of data over time.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://rasmusskytte.github.io/SCDB/}
}

}
\author{
\strong{Maintainer}: Rasmus Skytte Randløv \email{rske@ssi.dk} (\href{https://orcid.org/0000-0002-5860-3838}{ORCID})

Authors:
\itemize{
  \item Marcus Munch Grünewald \email{mmgr@ssi.dk} (\href{https://orcid.org/0009-0006-8090-406X}{ORCID})
}

}
\keyword{internal}
