% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCDB-package.R
\docType{package}
\name{SCDB-package}
\alias{SCDB}
\alias{SCDB-package}
\title{SCDB: Easily Access and Maintain Time-Based Versioned Data (Slowly-Changing-Dimension)}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A collection of functions that enable easy access and updating of a database of data over time. More specifically, the package facilitates type-2 history for data-warehouses and provides a number of Quality of life improvements for working on SQL databases with R. For reference see Ralph Kimball and Margy Ross (2013, ISBN 9781118530801).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ssi-dk/SCDB}
  \item Report bugs at \url{https://github.com/ssi-dk/SCDB/issues}
}

}
\author{
\strong{Maintainer}: Rasmus Skytte Randløv \email{rske@ssi.dk} (\href{https://orcid.org/0000-0002-5860-3838}{ORCID})

Authors:
\itemize{
  \item Marcus Munch Grünewald \email{mmgr@ssi.dk} (\href{https://orcid.org/0009-0006-8090-406X}{ORCID})
}

}
\keyword{internal}
