% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_accessing_functions.R, R/get_table.R
\name{table_exists}
\alias{table_exists}
\title{Check if table exists in db}
\usage{
table_exists(conn, db_table_id)

table_exists(conn, db_table_id)
}
\arguments{
\item{conn}{An object that inherits from DBIConnection (as generated by get_connection())'}

\item{db_table_id}{A \code{\link[DBI:Id]{DBI::Id()}} object or a character string readable by \link{id}.}
}
\description{
Check if table exists in db

Test if a table exists in database
}
\examples{
conn <- get_connection()

table_exists(conn, "prod.basis_samples") # TRUE
table_exists(conn, "data.basis_samples") # FALSE

close_connection(conn)
conn <- get_connection()

table_exists(conn, "prod.basis_samples") # TRUE
table_exists(conn, "data.basis_samples") # FALSE

close_connection(conn)
}
