% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scdb.r
\name{write_db}
\alias{write_db}
\title{Build database from object.}
\usage{
write_db(x, dbfile, layer = NULL, slurp = FALSE, verbose = TRUE)
}
\arguments{
\item{x}{object}

\item{dbfile}{SQLite file name}

\item{layer}{object layer name (defaults to the name of `x`)}

\item{verbose}{chatty mode}
}
\description{
Currently only objects supported by `spbabel::map_table` are supported ....
}

